--- Unity 5/Deprecated/readme.md
@@ -1,4 +1,0 @@
-Unity 5 to Babylon.js exporter
-==============================
-
-Documentation: http://doc.babylonjs.com/exporters/unity
