--- Unity 5/EditorToolkit/Redist/readme.md
@@ -1,28 +1,0 @@
-==========================================
-BabylonJS - Toolkit Binary Package
-==========================================
-
-Place the 'BabylonJS' folder containing the .unitypackage in your Unity Asset Store folder:
-
-- Mac: ~/Library/Unity/Asset Store-5x/
-
-- Windows: %AppData%\Roaming\Unity\Asset Store-5.x\
-
-- Example: ---> Asset Store-5.x
-             ---> BabylonJS
-                 ---> Babylon Editor Toolkit.unitypackage
-
-
-==========================================
-BabylonJS - Creating New Projects
-==========================================
-
-When creating a new unity project, select 'Add Asset Package' and enable 'BabylonJS Editor Toolkit' support.
-
-
-==========================================
-BabylonJS - Importing Existing Projects
-==========================================
-
-When using an existing unity project, select 'Assets->Import Package...' and choose your 'BabylonJS Editor Toolkit' unity package.
-
