--- Unity 5/EditorToolkit/readme.md
@@ -1,4 +1,0 @@
-Babylon.js EditorToolkit for Unity3D
-====================================
-
-Please visit our main documentation site: http://doc.babylonjs.com/resources/intro
