+++ Unity/EditorToolkit/readme.md
@@ -1,0 +1,4 @@
+Babylon.js EditorToolkit for Unity3D
+====================================
+
+Please visit our main documentation site: http://doc.babylonjs.com/resources/intro
